Move isCorePackage to utils
